circleci

+++ app/src/main/java/alirezamontazer/createbest/productcatalog_v3/Product.java
@@ -4,18 +4,18 @@ public class Product {
 
     private String _id;
     private String _productname;
-    private int _price;
+    private double _price;
 
     public Product() {
     }
 
-    public Product(String id, String productname, int price) {
+    public Product(String id, String productname, double price) {
         _id = id;
         _productname = productname;
         _price = price;
     }
 
-    public Product(String productname, int price) {
+    public Product(String productname, double price) {
         _productname = productname;
         _price = price;
     }
@@ -36,7 +36,7 @@ public class Product {
         return _productname;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         _price = price;
     }
